pyproject.toml
[project]
name = "21-human-in-the-loop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.14",
    "langgraph>=1.1.8",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
]
